New version 0.1.0.29; Include DDKBUILD_LOCAL.BAT (one path "above" cbm4win) if it exists for setting default values. This file is not controlled by CVS; Allow for DDKBUILD_ARGUMENTS (for example, for -prefast); Added building AMD64 and iA64 targets (by specifying -i386, -ia64, or -amd64 *after* -hollis); Removed inline assembly, added some casts around ULONG_PTR to make it compile on 64 bit systems (AMD64, iA64); Define ULONG_PTR for Linux so that libd64copy compiles on Linux; Use KeQueryPerformanceCounter() in perfeval.c if USE_TSC is not defined (for amd64, ia64).

+++ opencbm/include/cbmcopy.h
@@ -7,7 +7,7 @@
  *  Copyright 2001 Michael Klein <michael.klein@puffin.lb.shuttle.de>
 */
 
-/* $Id: cbmcopy.h,v 1.2 2005-07-21 17:39:20 strik Exp $ */
+/* $Id: cbmcopy.h,v 1.3 2005-09-11 13:32:33 strik Exp $ */
 
 #ifndef CBMCOPY_H
 #define CBMCOPY_H
@@ -76,7 +76,7 @@ extern int cbmcopy_read_file(CBM_FILE cbm_fd,
                              const unsigned char *cbmname,
                              int cbmname_size,
                              unsigned char **filedata,
-                             int *filedata_size,
+                             size_t *filedata_size,
                              cbmcopy_message_cb msg_cb,
                              cbmcopy_status_cb status_cb);
 
@@ -85,7 +85,7 @@ extern int cbmcopy_read_file_ts(CBM_FILE cbm_fd,
                                 int drive,
                                 int track, int sector,
                                 unsigned char **filedata,
-                                int *filedata_size,
+                                size_t *filedata_size,
                                 cbmcopy_message_cb msg_cb,
                                 cbmcopy_status_cb status_cb);
